:recycle: refactor(tags): Simplify implementation.
This is progress on #303 since we removed the use of withItem in favor
of hooks produced by makeItem. This is also progress on #184 since we
replace Meteor.call by api/call in AllergyCard (now
StaticAllergyCard). The hook builder `makeItem` is now properly typed.

These changes were prompted by ref forwarding issues with the custom
chips used in SetPicker.