:recycle: refactor(ui): Automated replace of makeStyles with styled.
Using npx @mui/codemod v5.0.0/jss-to-styled imports/. We had to undo
changes that broke transformed files. In particular this codemod
produces junk when using createStyles and ref, or when multiple
React.Fragment are used in the same file.

The codemod currently throws for CalendarData because we generate
styles on the fly there.

Some autogenerated styled components could be simplified when className
is only used on the root element.