:adhesive_bandage: fix(useReactive/useSubscription): StrictMode compat.
This should hopefully make all app tests pass now. This required
reimplementing both Meteor's useTracker and useSubscribe.

One call to react-fast-compare is necessary on first run of the effect
computation to avoid double rendering of the initial value.

Subscriptions seem to almost always reply with `#isReady() === false` on
first call, so we can/must put everything in the effect.