:closed_lock_with_key: security: Disable user-input eval in pdfjs-dist usage.
By default, pdfjs-dist optimizes some path resolution logic by compiling
a JavaScript function on the fly. The function is built using string
concatenation and no effort is made at sanitizing the parts it is
built from. These parts could contain user-input which leads to a code
injection vulnerability. This commit disables this default behavior.
An alternative is to upgrade pdfjs-dist to v4.2.67 or later.

For reference, see:
  - https://bugzilla.mozilla.org/show_bug.cgi?id=1893645
  - https://www.cve.org/CVERecord?id=CVE-2024-4367
  - https://security.snyk.io/vuln/SNYK-JS-PDFJSDIST-6810403
  - https://github.com/mozilla/pdf.js/security/advisories/GHSA-wgrm-67xf-hhpq
  - https://github.com/mozilla/pdf.js/pull/18015
  - https://github.com/wojtekmaj/react-pdf/discussions/1786
  - https://security.stackexchange.com/questions/248462/\
    is-firefoxs-new-javascript-support-within-pdf-files-a-security-concern/\
    248985
  - https://stackoverflow.com/questions/49299000/\
    what-are-the-security-implications-of-the-isevalsupported-option-in-pdf-js
  - https://github.com/mozilla/pdf.js/issues/10818

Not sure if this will break anything and/or will make certain things
slower.