:gear: config(ci): Increase `ci:build:compose` build timeout to 30 minutes.
We have to wait for the build to succeed to leverage the cache the
resulting updated cache. The build workflow might be sitting in the
queue for a long time before it actually start running when a lot of
concurrent workflows are running. Hence we increase the timeout to
decrease timeout failure rates.