:gear: config(ci): Configure restore key for puppeteer cache in test workflow.
This fixes #865 for good.

Since `meteor-actions/install` now can skip the installation of cached
packages using restore keys, it can happen that `node_modules/puppeteer`
is restored from cache while its `chromium` download is not.

With this commit, we ensure that the weakest restore key for `chromium`
is at least as weak as the weakest restore key for
`node_modules/puppeteer`.

Maybe an alternative would be to instruct `puppeteer` to install
chromium under `node_modules`? But for now, this is good enough.