:adhesive_bandage: fix(deploy): Allow to deploy with backup.
Before that a deployment with backup would fail because MongoDB only
bound to `patient-db` which can only be DNS-resolved once. By default,
resolution happens on the highest prio network which, by default, is the
first in alphabetical order. Hence it correctly bound to `patient-db`
inside `patient-backup-network`, but not in `patient-network`.

We now define and bind additional per-network hostnames for `patient-db`
(aliases).

An alternative would be to `--bind-ip 0.0.0.0`, since we control network
configuration anyway.